requirements.txt
# Add your project dependencies here, one per line.
# Example:
# requests
requests
python-dotenv
-e .
# numpy
